chore: update `core-lib-dev` app
